DRY MinterDAExp
