+++ test/minter-suite-minters/MinterDAExp/MinterDAExpV0.test.ts
@@ -6,9 +6,20 @@ import {
   balance,
   ether,
 } from "@openzeppelin/test-helpers";
+
 import { expect } from "chai";
 import { BigNumber } from "ethers";
 import { ethers } from "hardhat";
+
+import {
+  getAccounts,
+  assignDefaultConstants,
+  deployAndGet,
+  deployCoreWithMinterFilter,
+} from "../../util/common";
+import { ONE_MINUTE, ONE_HOUR, ONE_DAY } from "../../util/constants";
+import { MinterDAExp_Common } from "./MinterDAExp.common";
+
 import EthersAdapter from "@gnosis.pm/safe-ethers-lib";
 import Safe from "@gnosis.pm/safe-core-sdk";
 import { SafeTransactionDataPartial } from "@gnosis.pm/safe-core-sdk-types";
@@ -19,84 +30,60 @@ import { getGnosisSafe } from "../../util/GnosisSafeNetwork";
  * V1 core contract.
  */
 describe("MinterDAExpV0_V1Core", async function () {
-  const name = "Non Fungible Token";
-  const symbol = "NFT";
-
-  const firstTokenId = new BN("30000000");
-  const secondTokenId = new BN("3000001");
-
-  const startingPrice = ethers.utils.parseEther("10");
-  const higherPricePerTokenInWei = startingPrice.add(
-    ethers.utils.parseEther("0.1")
-  );
-  // purposefully different price per token on core contract (tracked separately)
-  const basePrice = ethers.utils.parseEther("0.05");
-
-  const projectOne = 3; // V1 core starts at project 3
-
-  const ONE_MINUTE = 60;
-  const ONE_HOUR = ONE_MINUTE * 60;
-  const ONE_DAY = ONE_HOUR * 24;
-
-  const defaultHalfLife = ONE_HOUR / 2;
-  const auctionStartTimeOffset = ONE_HOUR;
-
   beforeEach(async function () {
-    const [owner, newOwner, artist, additional, deployer] =
-      await ethers.getSigners();
-    this.accounts = {
-      owner: owner,
-      newOwner: newOwner,
-      artist: artist,
-      additional: additional,
-      deployer: deployer,
-    };
-
-    const randomizerFactory = await ethers.getContractFactory(
-      "BasicRandomizer"
+    // standard accounts and constants
+    this.accounts = await getAccounts.call(this);
+    await assignDefaultConstants.call(this, 3); // projectZero = 3 on V1 core
+    this.startingPrice = ethers.utils.parseEther("10");
+    this.higherPricePerTokenInWei = this.startingPrice.add(
+      ethers.utils.parseEther("0.1")
     );
-    this.randomizer = await randomizerFactory.deploy();
+    this.basePrice = ethers.utils.parseEther("0.05");
+    this.defaultHalfLife = ONE_HOUR / 2;
+    this.auctionStartTimeOffset = ONE_HOUR;
 
-    const artblocksFactory = await ethers.getContractFactory("GenArt721CoreV1");
-    this.token = await artblocksFactory
-      .connect(deployer)
-      .deploy(name, symbol, this.randomizer.address);
-
-    const minterFilterFactory = await ethers.getContractFactory(
+    // deploy and configure minter filter and minter
+    ({
+      genArt721Core: this.genArt721Core,
+      minterFilter: this.minterFilter,
+      randomizer: this.randomizer,
+    } = await deployCoreWithMinterFilter.call(
+      this,
+      "GenArt721CoreV1",
       "MinterFilterV0"
-    );
-    this.minterFilter = await minterFilterFactory.deploy(this.token.address);
+    ));
 
-    const minterFactory = await ethers.getContractFactory("MinterDAExpV0");
-    this.minter = await minterFactory.deploy(
-      this.token.address,
-      this.minterFilter.address
-    );
+    this.minter = await deployAndGet.call(this, "MinterDAExpV0", [
+      this.genArt721Core.address,
+      this.minterFilter.address,
+    ]);
 
-    await this.token
-      .connect(deployer)
+    await this.genArt721Core
+      .connect(this.accounts.deployer)
       .addProject("project1", this.accounts.artist.address, 0, false);
 
-    await this.token.connect(deployer).toggleProjectIsActive(projectOne);
+    await this.genArt721Core
+      .connect(this.accounts.deployer)
+      .toggleProjectIsActive(this.projectZero);
 
-    await this.token
-      .connect(deployer)
+    await this.genArt721Core
+      .connect(this.accounts.deployer)
       .addMintWhitelisted(this.minterFilter.address);
 
-    await this.token
-      .connect(artist)
-      .updateProjectMaxInvocations(projectOne, 15);
-
-    await this.token
+    await this.genArt721Core
       .connect(this.accounts.artist)
-      .toggleProjectIsPaused(projectOne);
+      .updateProjectMaxInvocations(this.projectZero, 15);
+
+    await this.genArt721Core
+      .connect(this.accounts.artist)
+      .toggleProjectIsPaused(this.projectZero);
 
     await this.minterFilter
       .connect(this.accounts.deployer)
       .addApprovedMinter(this.minter.address);
     await this.minterFilter
       .connect(this.accounts.deployer)
-      .setMinterForProject(projectOne, this.minter.address);
+      .setMinterForProject(this.projectZero, this.minter.address);
 
     if (!this.startTime) {
       const blockNumber = await ethers.provider.getBlockNumber();
@@ -108,140 +95,33 @@ describe("MinterDAExpV0_V1Core", async function () {
     await ethers.provider.send("evm_mine", [this.startTime - ONE_MINUTE]);
     await this.minter
       .connect(this.accounts.deployer)
-      .resetAuctionDetails(projectOne);
+      .resetAuctionDetails(this.projectZero);
     await this.minter
       .connect(this.accounts.artist)
       .setAuctionDetails(
-        projectOne,
-        this.startTime + auctionStartTimeOffset,
-        defaultHalfLife,
-        startingPrice,
-        basePrice
+        this.projectZero,
+        this.startTime + this.auctionStartTimeOffset,
+        this.defaultHalfLife,
+        this.startingPrice,
+        this.basePrice
       );
     await ethers.provider.send("evm_mine", [this.startTime]);
   });
 
-  describe("constructor", async function () {
-    it("reverts when given incorrect minter filter and core addresses", async function () {
-      const artblocksFactory = await ethers.getContractFactory(
-        "GenArt721CoreV3"
-      );
-      const token2 = await artblocksFactory
-        .connect(this.accounts.deployer)
-        .deploy(name, symbol, this.randomizer.address);
-
-      const minterFilterFactory = await ethers.getContractFactory(
-        "MinterFilterV0"
-      );
-      const minterFilter = await minterFilterFactory.deploy(token2.address);
-
-      const minterFactory = await ethers.getContractFactory(
-        "MinterSetPriceERC20V0"
-      );
-      // fails when combine new minterFilter with the old token in constructor
-      await expectRevert(
-        minterFactory.deploy(this.token.address, minterFilter.address),
-        "Illegal contract pairing"
-      );
-    });
-  });
-
-  describe("purchase", async function () {
-    it("disallows purchase before auction begins", async function () {
-      await ethers.provider.send("evm_mine", [this.startTime + ONE_HOUR / 2]);
-      await expectRevert(
-        this.minter.connect(this.accounts.owner).purchase(projectOne, {
-          value: startingPrice.toString(),
-          gasPrice: 0,
-        }),
-        "Auction not yet started"
-      );
-    });
-
-    it("calculates the price correctly", async function () {
-      for (let i = 1; i <= 5; i++) {
-        let ownerBalance = await this.accounts.owner.getBalance();
-        let price = startingPrice;
-        for (let j = 0; j < i; j++) {
-          price = price.div(2);
-        }
-
-        await ethers.provider.send("evm_setNextBlockTimestamp", [
-          this.startTime + auctionStartTimeOffset + i * defaultHalfLife,
-        ]);
-        await this.minter.connect(this.accounts.owner).purchase(projectOne, {
-          value: price.toString(),
-          gasPrice: 0,
-        });
-        // Test that price isn't too low
-
-        await expectRevert(
-          this.minter.connect(this.accounts.owner).purchase(projectOne, {
-            value: ((price.toBigInt() * BigInt(100)) / BigInt(101)).toString(),
-            gasPrice: 0,
-          }),
-          "Must send minimum value to mint!"
-        );
-        let ownerDelta = (await this.accounts.owner.getBalance()).sub(
-          ownerBalance
-        );
-        expect(ownerDelta.mul("-1").lte(price)).to.be.true;
-      }
-    });
-
-    it("calculates the price before correctly", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      await this.minter
-        .connect(this.accounts.artist)
-        .setAuctionDetails(
-          projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
-          startingPrice,
-          basePrice
-        );
-
-      let contractPriceInfo = await this.minter
-        .connect(this.accounts.owner)
-        .getPriceInfo(projectOne);
-      expect(contractPriceInfo.tokenPriceInWei).to.be.equal(startingPrice);
-    });
-
-    it("calculates the price after correctly ", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      await this.minter
-        .connect(this.accounts.artist)
-        .setAuctionDetails(
-          projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
-          startingPrice,
-          basePrice
-        );
-
-      await ethers.provider.send("evm_mine", [this.startTime + 5 * ONE_HOUR]);
-
-      let contractPriceInfo = await this.minter
-        .connect(this.accounts.owner)
-        .getPriceInfo(projectOne);
-      expect(contractPriceInfo.tokenPriceInWei).to.be.equal(basePrice);
-    });
+  describe("common tests", async function () {
+    MinterDAExp_Common();
   });
 
   describe("calculate gas", async function () {
     it("mints and calculates gas values", async function () {
       await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset,
+        this.startTime + this.auctionStartTimeOffset,
       ]);
 
       const tx = await this.minter
-        .connect(this.accounts.owner)
-        .purchase(projectOne, {
-          value: startingPrice,
+        .connect(this.accounts.user)
+        .purchase(this.projectZero, {
+          value: this.startingPrice,
         });
 
       const receipt = await ethers.provider.getTransactionReceipt(tx.hash);
@@ -259,35 +139,35 @@ describe("MinterDAExpV0_V1Core", async function () {
   describe("purchaseTo", async function () {
     it("allows `purchaseTo` by default", async function () {
       await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset,
+        this.startTime + this.auctionStartTimeOffset,
       ]);
       await this.minter
-        .connect(this.accounts.owner)
-        .purchaseTo(this.accounts.additional.address, projectOne, {
-          value: startingPrice,
+        .connect(this.accounts.user)
+        .purchaseTo(this.accounts.additional.address, this.projectZero, {
+          value: this.startingPrice,
         });
     });
 
     it("disallows `purchaseTo` if disallowed explicitly", async function () {
       await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset,
+        this.startTime + this.auctionStartTimeOffset,
       ]);
       await this.minter
         .connect(this.accounts.deployer)
-        .togglePurchaseToDisabled(projectOne);
+        .togglePurchaseToDisabled(this.projectZero);
       await expectRevert(
         this.minter
-          .connect(this.accounts.owner)
-          .purchaseTo(this.accounts.additional.address, projectOne, {
-            value: startingPrice,
+          .connect(this.accounts.user)
+          .purchaseTo(this.accounts.additional.address, this.projectZero, {
+            value: this.startingPrice,
           }),
         "No `purchaseTo` Allowed"
       );
       // still allows `purchaseTo` if destination matches sender.
       await this.minter
-        .connect(this.accounts.owner)
-        .purchaseTo(this.accounts.owner.address, projectOne, {
-          value: startingPrice,
+        .connect(this.accounts.user)
+        .purchaseTo(this.accounts.user.address, this.projectZero, {
+          value: this.startingPrice,
         });
     });
 
@@ -296,277 +176,18 @@ describe("MinterDAExpV0_V1Core", async function () {
       await expect(
         this.minter
           .connect(this.accounts.deployer)
-          .togglePurchaseToDisabled(projectOne)
+          .togglePurchaseToDisabled(this.projectZero)
       )
         .to.emit(this.minter, "PurchaseToDisabledUpdated")
-        .withArgs(projectOne, true);
+        .withArgs(this.projectZero, true);
       // emits false when changed from initial value of true
       await expect(
         this.minter
           .connect(this.accounts.deployer)
-          .togglePurchaseToDisabled(projectOne)
+          .togglePurchaseToDisabled(this.projectZero)
       )
         .to.emit(this.minter, "PurchaseToDisabledUpdated")
-        .withArgs(projectOne, false);
-    });
-  });
-
-  describe("setAuctionDetails", async function () {
-    it("cannot be modified mid-auction", async function () {
-      await ethers.provider.send("evm_mine", [
-        this.startTime + 2 * auctionStartTimeOffset,
-      ]);
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
-            startingPrice,
-            basePrice
-          ),
-        "No modifications mid-auction"
-      );
-    });
-
-    it("allows artist to set auction details", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      await this.minter
-        .connect(this.accounts.artist)
-        .setAuctionDetails(
-          projectOne,
-          this.startTime + auctionStartTimeOffset,
-          defaultHalfLife,
-          startingPrice,
-          basePrice
-        );
-    });
-
-    it("disallows whitelisted and non-artist to set auction details", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.additional)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
-            startingPrice,
-            basePrice
-          ),
-        "Only Artist"
-      );
-
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
-            startingPrice,
-            basePrice
-          ),
-        "Only Artist"
-      );
-    });
-
-    it("disallows higher resting price than starting price", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            defaultHalfLife,
-            basePrice,
-            startingPrice
-          ),
-        "Auction start price must be greater than auction end price"
-      );
-    });
-  });
-
-  describe("resetAuctionDetails", async function () {
-    it("allows whitelisted to reset auction details", async function () {
-      await expect(
-        this.minter
-          .connect(this.accounts.deployer)
-          .resetAuctionDetails(projectOne)
-      )
-        .to.emit(this.minter, "ResetAuctionDetails")
-        .withArgs(projectOne);
-    });
-
-    it("disallows artist to reset auction details", async function () {
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .resetAuctionDetails(projectOne),
-        "Only Core whitelisted"
-      );
-    });
-
-    it("disallows non-whitelisted non-artist to reset auction details", async function () {
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.additional)
-          .resetAuctionDetails(projectOne),
-        "Only Core whitelisted"
-      );
-    });
-
-    it("invalidates unpaused, ongoing auction (prevents price of zero)", async function () {
-      // prove projectOne is mintable
-      await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset,
-      ]);
-      await this.minter.connect(this.accounts.owner).purchase(projectOne, {
-        value: startingPrice,
-      });
-      // resetAuctionDetails for projectOne
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      // prove projectOne is no longer mintable
-      await expectRevert(
-        this.minter.connect(this.accounts.owner).purchase(projectOne, {
-          value: startingPrice,
-        }),
-        "Only configured auctions"
-      );
-      // prove projectOne is no longer mintable with zero value
-      // (always true given prior check, but paranoid so adding test)
-      await expectRevert(
-        this.minter.connect(this.accounts.owner).purchase(projectOne),
-        "Only configured auctions"
-      );
-    });
-  });
-
-  describe("enforce and broadcasts auction half-life", async function () {
-    it("enforces half-life min/max constraint", async function () {
-      await this.minter
-        .connect(this.accounts.deployer)
-        .resetAuctionDetails(projectOne);
-      // expect revert when creating a new project with
-      const invalidHalfLifeSecondsMin = ONE_MINUTE;
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            invalidHalfLifeSecondsMin,
-            startingPrice,
-            basePrice
-          ),
-        "Price decay half life must fall between min and max allowable values"
-      );
-
-      // expect revert when creating a new project with
-      const invalidHalfLifeSecondsMax = ONE_DAY;
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.artist)
-          .setAuctionDetails(
-            projectOne,
-            this.startTime + auctionStartTimeOffset,
-            invalidHalfLifeSecondsMax,
-            startingPrice,
-            basePrice
-          ),
-        "Price decay half life must fall between min and max allowable values"
-      );
-    });
-
-    it("emits event when allowable half life range is updated", async function () {
-      const newMinSeconds = 60;
-      const newMaxSeconds = 6000;
-      // emits event when allowable half life range is updated
-      await expect(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(
-            newMinSeconds,
-            newMaxSeconds
-          )
-      )
-        .to.emit(this.minter, "AuctionHalfLifeRangeSecondsUpdated")
-        .withArgs(newMinSeconds, newMaxSeconds);
-    });
-
-    it("validate setAllowablePriceDecayHalfLifeRangeSeconds guards", async function () {
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(600, 60),
-        "Maximum half life must be greater than minimum"
-      );
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.deployer)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(0, 600),
-        "Half life of zero not allowed"
-      );
-    });
-
-    it("validate setAllowablePriceDecayHalfLifeRangeSeconds ACL", async function () {
-      await expectRevert(
-        this.minter
-          .connect(this.accounts.additional)
-          .setAllowablePriceDecayHalfLifeRangeSeconds(60, 600),
-        "Only Core whitelisted"
-      );
-    });
-  });
-
-  describe("currency info hooks", async function () {
-    const unconfiguredProjectNumber = 99;
-
-    it("reports expected price per token", async function () {
-      // returns zero for unconfigured project price
-      const currencyInfo = await this.minter
-        .connect(this.accounts.artist)
-        .getPriceInfo(unconfiguredProjectNumber);
-      expect(currencyInfo.tokenPriceInWei).to.be.equal(0);
-    });
-
-    it("reports expected isConfigured", async function () {
-      let currencyInfo = await this.minter
-        .connect(this.accounts.artist)
-        .getPriceInfo(projectOne);
-      expect(currencyInfo.isConfigured).to.be.equal(true);
-      // false for unconfigured project
-      currencyInfo = await this.minter
-        .connect(this.accounts.artist)
-        .getPriceInfo(unconfiguredProjectNumber);
-      expect(currencyInfo.isConfigured).to.be.equal(false);
-    });
-
-    it("reports currency as ETH", async function () {
-      const priceInfo = await this.minter
-        .connect(this.accounts.artist)
-        .getPriceInfo(projectOne);
-      expect(priceInfo.currencySymbol).to.be.equal("ETH");
-    });
-
-    it("reports currency address as null address", async function () {
-      const priceInfo = await this.minter
-        .connect(this.accounts.artist)
-        .getPriceInfo(projectOne);
-      expect(priceInfo.currencyAddress).to.be.equal(constants.ZERO_ADDRESS);
+        .withArgs(this.projectZero, false);
     });
   });
 
@@ -575,10 +196,10 @@ describe("MinterDAExpV0_V1Core", async function () {
       // admin allows contract buys
       await this.minter
         .connect(this.accounts.deployer)
-        .toggleContractMintable(projectOne);
+        .toggleContractMintable(this.projectZero);
       // advance to time when auction is active
       await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset,
+        this.startTime + this.auctionStartTimeOffset,
       ]);
       // attacker deploys reentrancy contract
       const reentrancyMockFactory = await ethers.getContractFactory(
@@ -590,15 +211,15 @@ describe("MinterDAExpV0_V1Core", async function () {
       // attacker should see revert when performing reentrancy attack
       const totalTokensToMint = 2;
       let numTokensToMint = BigNumber.from(totalTokensToMint.toString());
-      let totalValue = higherPricePerTokenInWei.mul(numTokensToMint);
+      let totalValue = this.higherPricePerTokenInWei.mul(numTokensToMint);
       await expectRevert(
         reentrancyMock
           .connect(this.accounts.deployer)
           .attack(
             numTokensToMint,
             this.minter.address,
-            projectOne,
-            higherPricePerTokenInWei,
+            this.projectZero,
+            this.higherPricePerTokenInWei,
             {
               value: totalValue,
             }
@@ -608,17 +229,17 @@ describe("MinterDAExpV0_V1Core", async function () {
       );
       // attacker should be able to purchase ONE token at a time w/refunds
       numTokensToMint = BigNumber.from("1");
-      totalValue = higherPricePerTokenInWei.mul(numTokensToMint);
+      totalValue = this.higherPricePerTokenInWei.mul(numTokensToMint);
       for (let i = 0; i < totalTokensToMint; i++) {
         await reentrancyMock
           .connect(this.accounts.deployer)
           .attack(
             numTokensToMint,
             this.minter.address,
-            projectOne,
-            higherPricePerTokenInWei,
+            this.projectZero,
+            this.higherPricePerTokenInWei,
             {
-              value: higherPricePerTokenInWei,
+              value: this.higherPricePerTokenInWei,
             }
           );
       }
@@ -630,28 +251,28 @@ describe("MinterDAExpV0_V1Core", async function () {
       // admin allows contract buys
       await this.minter
         .connect(this.accounts.deployer)
-        .toggleContractMintable(projectOne);
+        .toggleContractMintable(this.projectZero);
       // advance to time when auction is active
       await ethers.provider.send("evm_mine", [
-        this.startTime + auctionStartTimeOffset,
+        this.startTime + this.auctionStartTimeOffset,
       ]);
 
       // deploy new Gnosis Safe
       const safeSdk: Safe = await getGnosisSafe(
         this.accounts.artist,
         this.accounts.additional,
-        this.accounts.owner
+        this.accounts.user
       );
       const safeAddress = safeSdk.getAddress();
 
       // create a transaction
       const unsignedTx = await this.minter.populateTransaction.purchase(
-        projectOne
+        this.projectZero
       );
       const transaction: SafeTransactionDataPartial = {
         to: this.minter.address,
         data: unsignedTx.data,
-        value: higherPricePerTokenInWei.toHexString(),
+        value: this.higherPricePerTokenInWei.toHexString(),
       };
       const safeTransaction = await safeSdk.createTransaction(transaction);
 
@@ -674,17 +295,17 @@ describe("MinterDAExpV0_V1Core", async function () {
       // fund the safe and execute transaction
       await this.accounts.artist.sendTransaction({
         to: safeAddress,
-        value: higherPricePerTokenInWei,
+        value: this.higherPricePerTokenInWei,
       });
-      const projectTokenInfoBefore = await this.token.projectTokenInfo(
-        projectOne
+      const projectTokenInfoBefore = await this.genArt721Core.projectTokenInfo(
+        this.projectZero
       );
       const executeTxResponse = await safeSdk2.executeTransaction(
         safeTransaction
       );
       await executeTxResponse.transactionResponse?.wait();
-      const projectTokenInfoAfter = await this.token.projectTokenInfo(
-        projectOne
+      const projectTokenInfoAfter = await this.genArt721Core.projectTokenInfo(
+        this.projectZero
       );
       expect(projectTokenInfoAfter.invocations).to.be.equal(
         projectTokenInfoBefore.invocations.add(1)
