+++ test/util/constants.ts
@@ -1,0 +1,3 @@
+export const ONE_MINUTE = 60;
+export const ONE_HOUR = ONE_MINUTE * 60;
+export const ONE_DAY = ONE_HOUR * 24;
